Title: 최대 수 만들기, Time: 1284ms, Memory: 62MB, Status: Passed - CodeTree

+++ 250323/최대 수 만들기/make-biggest-num.java
@@ -15,7 +15,7 @@ public class Main {
                 String strA = String.valueOf(a);
                 String strB = String.valueOf(b);
 
-                if (Integer.valueOf(strA + strB) > Integer.valueOf(strB + strA)) {
+                if (Long.valueOf(strA + strB) > Long.valueOf(strB + strA)) {
                     return -1;
                 } else {
                     return 1;
